Link sidebar and content editor
This commit allows the selecting of nodes in the sidebar to bring up that node's content in the content editor. The sidebar and the content editor are now linked, so that if no node is selected, the content editor is disabled, and when a node is selected, you can edit that node.